Move all users projection into the user package
